Generic CreateRow component
Separate the generic bits from the ink specific ones
