+++ app/javascript/src/admin/micro-clusters/DisplayMicroCluster.jsx
@@ -3,7 +3,9 @@ import React, { useContext } from "react";
 import { StateContext } from "./GenericApp";
 import { DisplayMacroClusters } from "./DisplayMacroClusters";
 import { CollectedInksList } from "./CollectedInksList";
-import { CreateRow } from "./CreateRow";
+import { CreateRow } from "../components/clustering/CreateRow";
+import { createMacroClusterAndAssign } from "./createMacroClusterAndAssign";
+import { ignoreCluster } from "./ignoreCluster";
 
 export const DisplayMicroCluster = ({ afterCreate, assignCluster }) => {
   const { activeCluster } = useContext(StateContext);
@@ -11,7 +13,12 @@ export const DisplayMicroCluster = ({ afterCreate, assignCluster }) => {
     <div className="fpc-table fpc-table--full-width fpc-scroll-shadow">
       <table className="table">
         <thead>
-          <CreateRow afterCreate={afterCreate} assignCluster={assignCluster} />
+          <CreateRow
+            afterCreate={afterCreate}
+            createMacroClusterAndAssign={createMacroClusterAndAssign}
+            ignoreCluster={ignoreCluster}
+            fields={["brand_name", "line_name", "ink_name"]}
+          />
         </thead>
         <tbody>
           <CollectedInksList collectedInks={activeCluster.entries} />
